[문제 해결] 입력하여 바로 검색 기능 관련 문제 해결 * 검색 시 프로그램이 멈추는 문제 해결 * 글자 수 슬라이더의 숫자가 이제는 정수로 나타납니다.

+++ lib/open_class.dart
@@ -53,7 +53,7 @@ class _OpenClassState extends State<OpenClass> {
   List orgClassList = [];
   bool _offline = false;
   bool _liveSearch = true;
-  int _liveSearchCount = 0;
+  double _liveSearchCount = 0.0;
   bool _isSaved = false;
   String _myDept = '컴퓨터학부';
   String _mySub = '학부 공통';
@@ -79,7 +79,7 @@ class _OpenClassState extends State<OpenClass> {
                 true;
         _liveSearchCount = (jsonDecode(_pref.getString('settings')!)
                 as Map)['liveSearchCount'] ??
-            0;
+            0.0;
       }
     }
     if ((_pref.containsKey('db_ver')) && _offline) {
